Prevent integer multiply overflow in calls to cram_not_enough_bits
An example in cram_beta_decode_int:  if we have out_size as 134217729
and nbits as 32 then we overflow and the call to cram_not_enough_bits
passes provided our input buffer is >= 32. This is because int32_t
134217729*32 gives 32 due to overflow.  Casting one of the values to
uint64_t first gives us the full size and we fail the input buffer
size validation.

This adjusts the cram_not_enough_bits() interface and calls
to it from the BETA codec.

(cherry picked from commit 3eef4aece30bc008b251e377085fc944d008d06a)

+++ cram/cram_codecs.c
@@ -1071,10 +1071,10 @@ cram_codec *cram_const_encode_init(cram_stats *st,
  */
 int cram_beta_decode_long(cram_slice *slice, cram_codec *c, cram_block *in, char *out, int *out_size) {
     int64_t *out_i = (int64_t *)out;
-    int i, n = *out_size;
+    uint32_t i, n = *out_size;
 
     if (c->u.beta.nbits) {
-        if (cram_not_enough_bits(in, c->u.beta.nbits * n))
+        if (cram_not_enough_bits(in, c->u.beta.nbits * (uint64_t)n))
             return -1;
 
         for (i = 0; i < n; i++)
@@ -1089,10 +1089,10 @@ int cram_beta_decode_long(cram_slice *slice, cram_codec *c, cram_block *in, char
 
 int cram_beta_decode_int(cram_slice *slice, cram_codec *c, cram_block *in, char *out, int *out_size) {
     int32_t *out_i = (int32_t *)out;
-    int i, n = *out_size;
+    uint32_t i, n = *out_size;
 
     if (c->u.beta.nbits) {
-        if (cram_not_enough_bits(in, c->u.beta.nbits * n))
+        if (cram_not_enough_bits(in, c->u.beta.nbits * (uint64_t)n))
             return -1;
 
         for (i = 0; i < n; i++)
@@ -1106,11 +1106,11 @@ int cram_beta_decode_int(cram_slice *slice, cram_codec *c, cram_block *in, char 
 }
 
 int cram_beta_decode_char(cram_slice *slice, cram_codec *c, cram_block *in, char *out, int *out_size) {
-    int i, n = *out_size;
+    uint32_t i, n = *out_size;
 
 
     if (c->u.beta.nbits) {
-        if (cram_not_enough_bits(in, c->u.beta.nbits * n))
+        if (cram_not_enough_bits(in, c->u.beta.nbits * (uint64_t)n))
             return -1;
 
         if (out)
